break[array]: move like compute to lazy compute
