+++ vortex-array/src/arrays/dict/compute/mod.rs
@@ -6,7 +6,6 @@ mod compare;
 mod fill_null;
 mod is_constant;
 mod is_sorted;
-mod like;
 mod min_max;
 pub(crate) mod rules;
 mod slice;
