+++ vortex-array/src/compute/mod.rs
@@ -27,7 +27,6 @@ pub use invert::invert;
 pub use is_constant::*;
 pub use is_sorted::*;
 use itertools::Itertools;
-pub use like::*;
 pub use list_contains::*;
 pub use mask::*;
 pub use min_max::*;
@@ -73,7 +72,6 @@ mod filter;
 mod invert;
 mod is_constant;
 mod is_sorted;
-mod like;
 mod list_contains;
 mod mask;
 mod min_max;
@@ -99,7 +97,6 @@ pub fn warm_up_vtables() {
     compare::warm_up_vtable();
     is_constant::warm_up_vtable();
     is_sorted::warm_up_vtable();
-    like::warm_up_vtable();
     list_contains::warm_up_vtable();
     mask::warm_up_vtable();
     min_max::warm_up_vtable();
