+++ vortex-datafusion/src/convert/exprs.rs
@@ -19,13 +19,13 @@ use datafusion_physical_expr::utils::collect_columns;
 use datafusion_physical_expr_common::physical_expr::is_dynamic_physical_expr;
 use datafusion_physical_plan::expressions as df_expr;
 use itertools::Itertools;
-use vortex::compute::LikeOptions;
 use vortex::dtype::DType;
 use vortex::dtype::Nullability;
 use vortex::dtype::arrow::FromArrowType;
 use vortex::expr::Binary;
 use vortex::expr::Expression;
 use vortex::expr::Like;
+use vortex::expr::LikeOptions;
 use vortex::expr::Operator;
 use vortex::expr::VTableExt;
 use vortex::expr::and_collect;
