+++ vortex-duckdb/src/convert/expr.rs
@@ -6,7 +6,6 @@ use std::sync::Arc;
 use itertools::Itertools;
 use tracing::debug;
 use vortex::compute::BetweenOptions;
-use vortex::compute::LikeOptions;
 use vortex::compute::StrictComparison;
 use vortex::dtype::Nullability;
 use vortex::error::VortexError;
@@ -18,6 +17,7 @@ use vortex::expr::Between;
 use vortex::expr::Binary;
 use vortex::expr::Expression;
 use vortex::expr::Like;
+use vortex::expr::LikeOptions;
 use vortex::expr::Literal;
 use vortex::expr::Operator;
 use vortex::expr::VTableExt;
